prepare structure for additional kinematic_extensions
