+++ cob_twist_controller/include/cob_twist_controller/inverse_differential_kinematics_solver.h
@@ -36,7 +36,7 @@
 #include "cob_twist_controller/cob_twist_controller_data_types.h"
 #include "cob_twist_controller/callback_data_mediator.h"
 #include "cob_twist_controller/limiters/limiter.h"
-#include "cob_twist_controller/kinematic_extensions/kinematic_extension.h"
+#include "cob_twist_controller/kinematic_extensions/kinematic_extension_builder.h"
 #include "cob_twist_controller/constraint_solvers/constraint_solver_factory.h"
 #include "cob_twist_controller/task_stack/task_stack_controller.h"
