+++ cob_twist_controller/src/kinematic_extensions/kinematic_extension_urdf.cpp
@@ -1,0 +1,32 @@
+/*!
+ *****************************************************************
+ * \file
+ *
+ * \note
+ *   Copyright (c) 2015 \n
+ *   Fraunhofer Institute for Manufacturing Engineering
+ *   and Automation (IPA) \n\n
+ *
+ *****************************************************************
+ *
+ * \note
+ *   Project name: care-o-bot
+ * \note
+ *   ROS stack name: cob_control
+ * \note
+ *   ROS package name: cob_twist_controller
+ *
+ * \author
+ *   Author: Felix Messmer, email: felix.messmer@ipa.fraunhofer.de
+ *
+ * \date Date of creation: June, 2015
+ *
+ * \brief
+ *   This header contains the interface description for extening the
+ *   kinematic chain with additional degrees of freedom, e.g. base_active or lookat
+ *
+ ****************************************************************/
+
+#include <eigen_conversions/eigen_kdl.h>
+#include "cob_twist_controller/kinematic_extensions/kinematic_extension.h"
+
